feat: delete legacy code and organize new code
