+++ src/index.ts
@@ -1,2 +1,1 @@
-export { serializeFunction } from "./closure/serializeClosure";
-export { serializeFunction as serializeFunction2 } from "./closure/serializeClosure2";
+export { serializeFunction } from "./serialize-closure";
