--- src/log.ts
@@ -1,44 +1,0 @@
-// Copyright 2016-2018, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * debug logs a debug-level message that is generally hidden from end-users.
- */
-export function debug(_msg: string) {
-  return Promise.resolve();
-}
-
-/**
- * info logs an informational message that is generally printed to stdout during resource operations.
- */
-export function info(msg: string) {
-  console.log(`info: [runtime] ${msg}`);
-  return Promise.resolve();
-}
-
-/**
- * warn logs a warning to indicate that something went wrong, but not catastrophically so.
- */
-export function warn(msg: string) {
-  console.warn(`warning: [runtime] ${msg}`);
-  return Promise.resolve();
-}
-
-/**
- * error logs a fatal condition. Consider raising an exception after calling error to stop the Pulumi program.
- */
-export function error(msg: string) {
-  console.error(`error: [runtime] ${msg}`);
-  return Promise.resolve();
-}
